Got basic 2d object construction down, so far just renders a little object that bounces around

+++ src/types.h
@@ -6,15 +6,15 @@
 #define TYPES_H
 
 struct Rotation {
-    double pitch;
-    double roll;
-    double yaw;
+    float pitch;
+    float roll;
+    float yaw;
 };
 
 struct Position {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 
     Rotation rotation;
 };
